pyproject.toml
[tool.poetry]
name = "file-finder"
version = "1.0.0"
description = "A Python module for finding files"
authors = ["Tyler Fisher <tylerfisher@tylerfisher.ca>"]
readme = "README.md"
packages = [{include = "file_finder"}]

[tool.poetry.dependencies]
python = "<3.12,>=3.7"
click = "^8.1.3"
yara-python = "^4.2.3"
tqdm = "^4.65.0"
pyinstaller = "^5.9.0"
pytest = "^7.2.2"
humanize = "^4.6.0"

[tool.poetry.scripts]
find = "file_finder.cli:main"
search = "file_finder.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
